+++ CHANGELOG.md
@@ -4,7 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
+### ✅ Added
+- Improved support for moderation events handling. [#1004](https://github.com/GetStream/stream-video-swift/pull/1004)
 
 # [1.37.0](https://github.com/GetStream/stream-video-swift/releases/tag/1.37.0)
 _November 28, 2025_